requirements.txt
mysql-connector-python==9.3.0
numpy==2.3.1
pandas==2.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
six==1.17.0
tzdata==2025.2
